+++ classes/CorrectorAdmin/class.CorrectorAdminGUI.php
@@ -30,17 +30,15 @@ use ILIAS\DI\Exceptions\Exception;
  */
 class CorrectorAdminGUI extends BaseGUI
 {
-
-    /** @var CorrectorAdminService */
-    protected $service;
-    
-    /** @var CorrectionSettings */
-    protected $settings;
+    protected CorrectorAdminService $service;
+    protected CorrectorAssignmentsService  $assignment_service;
+    protected CorrectionSettings $settings;
 
     public function __construct(\ilObjLongEssayAssessmentGUI $objectGUI)
     {
         parent::__construct($objectGUI);
         $this->service = $this->localDI->getCorrectorAdminService($this->object->getId());
+        $this->assignment_service = $this->localDI->getCorrectorAssignmentService($this->object->getId());
         $this->settings = $this->localDI->getTaskRepo()->getCorrectionSettingsById($this->object->getId());
     }
 
@@ -696,8 +694,6 @@ class CorrectorAdminGUI extends BaseGUI
 
     public function correctorAssignmentSpreadsheetImport()
     {
-        $corrector_repo = $this->localDI->getCorrectorRepo();
-        $task_repo = $this->localDI->getTaskRepo();
         $tempfile = new ilLongEssayAssessmentUploadTempFile($this->storage, $this->dic->filesystem(), $this->dic->upload());
 
         $form = $this->uiFactory->input()->container()->form()->standard(
@@ -707,7 +703,8 @@ class CorrectorAdminGUI extends BaseGUI
                     $this->storage,
                     $tempfile
                 ),
-                $this->plugin->txt("assignment_excel_import")
+                $this->plugin->txt("assignment_excel_import"),
+                $this->plugin->txt("assignment_excel_import_info")
             )]
         );
 
@@ -716,172 +713,36 @@ class CorrectorAdminGUI extends BaseGUI
 
             if($data = $form->getData()) {
                 $filename = $data['excel'][0];
-                $needed_corr = $task_repo->getCorrectionSettingsById($this->object->getId())->getRequiredCorrectors();
 
                 try {
-                    $spreadsheet = new CorrectorAssignmentExcel();
-                    $spreadsheet->loadFromFile(\ilUtil::getDataDir() . '/temp/' . $filename);
-
-                    $spreadsheet->setActiveSheet(1);
-                    $corrector = [];
-                    $r = 2;
-                    while (($id = $spreadsheet->getCell($r, 0)) != null) {
-                        $login = $spreadsheet->getCell($r, 1);
-                        $corrector[$login] = $id;
-                        $r++;
-                    }
-                    $spreadsheet->setActiveSheet(0);
-                    $corrector_assignments = [];
-                    $r = 2;
-                    while (($id = $spreadsheet->getCell($r, 0)) != null) {
-                        $assigned = [];
-                        foreach(range(0, $needed_corr-1) as $pos) {
-                            $login = $spreadsheet->getCell($r, 8+$pos);
-                            $assigned[$pos] = (int) $corrector[$login] ?? CorrectorAdminService::BLANK_CORRECTOR_ASSIGNMENT;
-                        }
-
-                        $corrector_assignments[$id] = $assigned;
-                        $r++;
-                    }
-
-                    foreach($corrector_assignments as $writer_id => $as) {
-                        $fa = $as[0];
-                        $sa = $needed_corr == 2 ? $as[1] : CorrectorAdminService::UNCHANGED_CORRECTOR_ASSIGNMENT;
-
-                        $this->service->assignMultipleCorrector($fa, $sa, [$writer_id]);
-                    }
+                    $this->assignment_service->importAssignments(\ilUtil::getDataDir() . '/temp/' . $filename);
+                    ilUtil::sendSuccess($this->plugin->txt("corrector_assignment_change_file_success"), true);
                     $tempfile->removeTempFile($filename);
-                    ilUtil::sendSuccess($this->plugin->txt("corrector_assignment_changed"), true);
                     $this->ctrl->redirect($this);
-                } catch (\Exception $exception) {
+                }
+                catch (CorrectorAssignmentsException $exception) {
                     $tempfile->removeTempFile($filename);
-                    ilUtil::sendFailure($this->plugin->txt("corrector_assignment_change_file_failure"), true);
+                    ilUtil::sendFailure($this->plugin->txt("corrector_assignment_change_file_failure")
+                        . '<p class="small">' . nl2br($exception->getMessage()), true) . '</p>';
+                }
+                catch (\Exception $exception) {
+                    $tempfile->removeTempFile($filename);
+                    ilUtil::sendFailure($this->plugin->txt("corrector_assignment_change_file_failure"));
                 }
             }
         }
-        ilUtil::sendInfo($this->plugin->txt("change_corrector_info"));
+
         $this->tpl->setContent($this->renderer->render($form));
     }
 
     public function correctorAssignmentSpreadsheetExport()
     {
-        $corrector_repo = $this->localDI->getCorrectorRepo();
-        $writer_repo = $this->localDI->getWriterRepo();
-        $task_repo = $this->localDI->getTaskRepo();
-        $essay_repo = $this->localDI->getEssayRepo();
-        $needed_corr = $task_repo->getCorrectionSettingsById($this->object->getId())->getRequiredCorrectors();
-        $participant_title = "Writer";
-        $corrector_title = "Corrector";
-        $locations = [];
-
-        foreach($task_repo->getLocationsByTaskId($this->object->getId()) as $location) {
-            $locations[$location->getId()] = $location;
+        try {
+            $this->assignment_service->exportAssignments($this->spreadsheetAssignmentToggle());
         }
-
-        $essays = [];
-
-        foreach($essay_repo->getEssaysByTaskId($this->object->getId()) as $essay) {
-            $essays[$essay->getWriterId()] = $essay;
+        catch (Exception $e) {
+            // maybe logging
         }
-        $corrector = [];
-        foreach($corrector_repo->getCorrectorsByTaskId($this->object->getId()) as $r) {
-            $corrector[$r->getId()] = $r;
-        }
-
-        $writer = $writer_repo->getWritersByTaskId($this->object->getId());
-
-        if($this->spreadsheetAssignmentToggle()) {
-
-            $authorized_user = [];
-            foreach($essays as $writer_id => $essay) {
-                if($essay->getWritingAuthorized() !== null) {
-                    $authorized_user[] = $essay->getWriterId();
-                }
-            }
-            $writer = array_filter($writer, fn ($x) => in_array($x->getId(), $authorized_user));
-        }
-
-        $users = [];
-
-        foreach(\ilObjUser::_getUserData(array_merge(
-            array_map(fn (Corrector $x) => $x->getUserId(), $corrector),
-            array_map(fn (Writer $x) => $x->getUserId(), $writer),
-        )) as $u) {
-            $users[(int)$u['usr_id']] = $u;
-        }
-
-        $assignments = [];
-
-        foreach($corrector_repo->getAssignmentsByTaskId($this->object->getId()) as $assignment) {
-            $assignments[$assignment->getWriterId()][$assignment->getPosition()] = $assignment;
-        }
-        $r = 2;
-        $spreadsheet = new CorrectorAssignmentExcel();
-        $writer_sheet = $spreadsheet->addSheet($participant_title, true);
-        $corrector_sheet = $spreadsheet->addSheet($corrector_title, false);
-        $spreadsheet->setCell(1, 0, 'Id');
-        $spreadsheet->setCell(1, 1, 'Login');
-        $spreadsheet->setCell(1, 2, 'Firstname');
-        $spreadsheet->setCell(1, 3, 'Lastname');
-        $spreadsheet->setCell(1, 4, 'Email');
-        $spreadsheet->setCell(1, 5, 'Pseudonym');
-        $spreadsheet->setCell(1, 6, 'Location');
-        $spreadsheet->setCell(1, 7, 'Words');
-        foreach(range(0, $needed_corr-1) as $pos) {
-            $spreadsheet->setCell(1, 8+$pos, 'Corrector ' . ($pos+1));
-        }
-
-        foreach($writer as $w) {
-            $data = $users[$w->getUserId()] ?? [];
-            $ass = $assignments[$w->getId()] ?? [];
-            ksort($ass);
-            $essay = $essays[$w->getId()] ?? null;
-            $location = $essay !== null ? $locations[$essay->getLocation()] ?? null : null;
-            $written_text = $essay !== null ? $essay->getWrittenText() : "";
-            $location_text = $location !== null ? $location->getTitle() : "";
-
-            $spreadsheet->setCell($r, 0, $w->getId());
-            $spreadsheet->setCell($r, 1, $data['login'] ?? "");
-            $spreadsheet->setCell($r, 2, $data['firstname'] ?? "");
-            $spreadsheet->setCell($r, 3, $data['lastname'] ?? "");
-            $spreadsheet->setCell($r, 4, $data['email'] ?? "");
-            $spreadsheet->setCell($r, 5, $w->getPseudonym());
-            $spreadsheet->setCell($r, 6, $location_text);
-            $spreadsheet->setCell($r, 7, str_word_count($written_text));
-
-            foreach(range(0, $needed_corr-1) as $pos) {
-                $spreadsheet->addDropdownCol($r, 8+$pos, '=\''.$corrector_title.'\'!$B$2:$B$'.(count($corrector)+1));
-            }
-
-            foreach ($ass as $a) {
-                $c = $corrector[$a->getCorrectorId()] ?? null;
-                $login = $c !== null && isset($users[$c->getUserId()]) ? $users[$c->getUserId()]['login'] ?? '' : '';
-                $spreadsheet->setCell($r, 8+$a->getPosition(), $login);
-            }
-            $r++;
-        }
-        $r = 2;
-        $spreadsheet->setActiveSheet($corrector_sheet);
-        $spreadsheet->setCell(1, 0, "Id");
-        $spreadsheet->setCell(1, 1, 'Login');
-        $spreadsheet->setCell(1, 2, 'Firstname');
-        $spreadsheet->setCell(1, 3, 'Lastname');
-        $spreadsheet->setCell(1, 4, 'Email');
-
-        foreach($corrector as $c) {
-            $data = $users[$c->getUserId()] ?? [];
-            $ass = $assignments[$c->getId()] ?? [];
-            ksort($ass);
-
-            $spreadsheet->setCell($r, 0, $c->getId());
-            $spreadsheet->setCell($r, 1, $data['login'] ?? "");
-            $spreadsheet->setCell($r, 2, $data['firstname'] ?? "");
-            $spreadsheet->setCell($r, 3, $data['lastname'] ?? "");
-            $spreadsheet->setCell($r, 4, $data['email'] ?? "");
-            $r++;
-        }
-        $spreadsheet->setActiveSheet($corrector_sheet);
-        $spreadsheet->sendToClient("corrector_assignment.xlsx");
         $this->ctrl->redirect($this);
     }